Add mouseout event on frame to hide tooltip
When the tooltip was visible on a data point very close to the top,
the tooltip would not hide when moving the mouse out of the frame.

Now, the tooltip always hides when the mouse goes out of the frame.
It uses a parent-check to simulate the `mouseleave` event (which is only
available on IE).