Added floatingRange type to the timeline
I have added a floatingRange type to this timeline.
This new type is like the range type but the startDate and endDate are both optional.

Therefor a range with only a startDate will have no definite endDate it will carry on till the end of the visible timeline.
A range with only an endDate will start at the beginning of the visible timeline and carry on till the endDate.
A given range with no startDate and no endDate will start at the beginning of the visible timeline and carry on till the end of the visible timeline.
The drag and drop handles are still working and they will set the start and endDate accordingly when dragged.
The delete image will no be visible when the endDate is not set. You must drag the endDate draghandle to be able to delete the item. (ideas for a better solution are welcome)

Enjoy.